refactor(infrastructure): refactor infrastructure folder structure to domain based organization (checkout, cart)
