+++ apps/storefront/src/services/index.ts
@@ -1,10 +1,10 @@
 import * as Sentry from "@sentry/nextjs";
 
+import { saleorCartService } from "@nimara/infrastructure/cart/index";
+import { saleorCheckoutService } from "@nimara/infrastructure/checkout/service";
 import type { ErrorService } from "@nimara/infrastructure/error/service";
 import { saleorAddressService } from "@nimara/infrastructure/public/saleor/address/index";
 import { saleorAuthService } from "@nimara/infrastructure/public/saleor/auth/index";
-import { saleorCartService } from "@nimara/infrastructure/public/saleor/cart/index";
-import { saleorCheckoutService } from "@nimara/infrastructure/public/saleor/checkout/service";
 import { saleorCollectionService } from "@nimara/infrastructure/public/saleor/collection/index";
 import { saleorFulfillmentService } from "@nimara/infrastructure/public/saleor/fulfillment/service";
 import { saleorStoreService } from "@nimara/infrastructure/public/saleor/store/index";
