+++ packages/infrastructure/src/cart/providers.ts
@@ -3,11 +3,11 @@ import { linesAddUseCase } from "#root/use-cases/cart/lines-add-use-case";
 import { linesDeleteUseCase } from "#root/use-cases/cart/lines-delete-use-case";
 import { linesUpdateUseCase } from "#root/use-cases/cart/lines-update-use-case";
 
-import { saleorCartCreateInfra } from "./infrastructure/cart-create-infra";
-import { saleorCartGetInfra } from "./infrastructure/cart-get-infra";
-import { saleorLinesAddInfra } from "./infrastructure/lines-add-infra";
-import { saleorLinesDeleteInfra } from "./infrastructure/lines-delete-infra";
-import { saleorLinesUpdateInfra } from "./infrastructure/lines-update-infra";
+import { saleorCartCreateInfra } from "./saleor/infrastructure/cart-create-infra";
+import { saleorCartGetInfra } from "./saleor/infrastructure/cart-get-infra";
+import { saleorLinesAddInfra } from "./saleor/infrastructure/lines-add-infra";
+import { saleorLinesDeleteInfra } from "./saleor/infrastructure/lines-delete-infra";
+import { saleorLinesUpdateInfra } from "./saleor/infrastructure/lines-update-infra";
 import type { CartService, SaleorCartServiceConfig } from "./types";
 
 export const saleorCartService: CartService<SaleorCartServiceConfig> = (
