+++ packages/infrastructure/src/cart/saleor/graphql/mutations/CartLinesDeleteMutation.graphql
