+++ packages/infrastructure/src/cart/saleor/infrastructure/lines-delete-infra.ts
@@ -2,8 +2,8 @@ import { err, ok } from "@nimara/domain/objects/Result";
 
 import { graphqlClient } from "#root/graphql/client";
 
+import type { LinesDeleteInfra, SaleorCartServiceConfig } from "../../types";
 import { CartLinesDeleteMutationDocument } from "../graphql/mutations/generated";
-import type { LinesDeleteInfra, SaleorCartServiceConfig } from "../types";
 
 export const saleorLinesDeleteInfra =
   ({ apiURI, logger }: SaleorCartServiceConfig): LinesDeleteInfra =>
