+++ packages/infrastructure/src/checkout/saleor/graphql/fragments/CheckoutErrorFragment.graphql
