+++ packages/infrastructure/src/checkout/saleor/graphql/fragments/CheckoutFragment.graphql
