+++ packages/infrastructure/src/checkout/saleor/graphql/mutations/CheckoutCustomerAttachMutation.graphql
