+++ packages/infrastructure/src/checkout/saleor/graphql/mutations/CheckoutRemovePromoCodeMutation.graphql
