+++ packages/infrastructure/src/checkout/saleor/infrastructure/checkout-add-promo-code-infra.ts
@@ -3,11 +3,11 @@ import { err, ok } from "@nimara/domain/objects/Result";
 import { graphqlClient } from "#root/graphql/client";
 import { handleMutationErrors } from "#root/public/saleor/error";
 
-import { CheckoutAddPromoCodeMutationDocument } from "../graphql/mutations/generated";
 import type {
   CheckoutAddPromoCodeInfra,
   SaleorCheckoutServiceConfig,
-} from "../types";
+} from "../../types";
+import { CheckoutAddPromoCodeMutationDocument } from "../graphql/mutations/generated";
 
 export const saleorCheckoutAddPromoCodeInfra =
   ({
