+++ packages/infrastructure/src/checkout/saleor/infrastructure/checkout-remove-promo-code-infra.ts
@@ -3,11 +3,11 @@ import { err, ok } from "@nimara/domain/objects/Result";
 import { graphqlClient } from "#root/graphql/client";
 import { handleMutationErrors } from "#root/public/saleor/error";
 
-import { CheckoutRemovePromoCodeMutationDocument } from "../graphql/mutations/generated";
 import type {
   CheckoutRemovePromoCodeInfra,
   SaleorCheckoutServiceConfig,
-} from "../types";
+} from "../../types";
+import { CheckoutRemovePromoCodeMutationDocument } from "../graphql/mutations/generated";
 
 export const saleorCheckoutRemovePromoCodeInfra =
   ({
