+++ packages/infrastructure/src/checkout/service.ts
@@ -8,15 +8,15 @@ import { checkoutRemovePromoCodeUseCase } from "#root/use-cases/checkout/checkou
 import { checkoutShippingAddressUpdateUseCase } from "#root/use-cases/checkout/checkout-shipping-address-update-use-case";
 import { orderCreateUseCase } from "#root/use-cases/checkout/order-create-use-case";
 
-import { saleorCheckoutAddPromoCodeInfra } from "./infrastructure/checkout-add-promo-code-infra";
-import { saleorCheckoutCustomerAttachInfra } from "./infrastructure/checkout-customer-attach-infrastructure";
-import { saleorDeliveryMethodUpdateInfra } from "./infrastructure/checkout-delivery-method-update-infra";
-import { saleorCheckoutGetInfra } from "./infrastructure/checkout-get-infra";
-import { saleorCheckoutRemovePromoCodeInfra } from "./infrastructure/checkout-remove-promo-code-infra";
-import { saleorCheckoutShippingAddressUpdateInfra } from "./infrastructure/checkout-shipping-address-update-infra";
-import { saleorCheckoutEmailUpdateInfra } from "./infrastructure/checkout-update-email-infra";
-import { orderCreateInfra } from "./infrastructure/order-create-infra";
-import { saleorCheckoutBillingAddressUpdateInfra } from "./infrastructure/update-checkout-billing-address-infra";
+import { saleorCheckoutAddPromoCodeInfra } from "./saleor/infrastructure/checkout-add-promo-code-infra";
+import { saleorCheckoutCustomerAttachInfra } from "./saleor/infrastructure/checkout-customer-attach-infrastructure";
+import { saleorDeliveryMethodUpdateInfra } from "./saleor/infrastructure/checkout-delivery-method-update-infra";
+import { saleorCheckoutGetInfra } from "./saleor/infrastructure/checkout-get-infra";
+import { saleorCheckoutRemovePromoCodeInfra } from "./saleor/infrastructure/checkout-remove-promo-code-infra";
+import { saleorCheckoutShippingAddressUpdateInfra } from "./saleor/infrastructure/checkout-shipping-address-update-infra";
+import { saleorCheckoutEmailUpdateInfra } from "./saleor/infrastructure/checkout-update-email-infra";
+import { orderCreateInfra } from "./saleor/infrastructure/order-create-infra";
+import { saleorCheckoutBillingAddressUpdateInfra } from "./saleor/infrastructure/update-checkout-billing-address-infra";
 import type { CheckoutService, SaleorCheckoutServiceConfig } from "./types";
 
 export const saleorCheckoutService: CheckoutService<
