+++ packages/infrastructure/src/use-cases/checkout/checkout-billing-address-update-use-case.ts
@@ -1,7 +1,7 @@
 import type {
   CheckoutBillingAddressUpdateInfra,
   CheckoutBillingAddressUpdateUseCase,
-} from "#root/public/saleor/checkout/types";
+} from "#root/checkout/types";
 
 export const checkoutBillingAddressUpdateUseCase = ({
   checkoutBillingAddressUpdateInfra,
