+++ packages/infrastructure/src/use-cases/checkout/checkout-remove-promo-code-use-case.ts
@@ -1,7 +1,7 @@
 import type {
   CheckoutRemovePromoCodeInfra,
   CheckoutRemovePromoCodeUseCase,
-} from "#root/public/saleor/checkout/types";
+} from "#root/checkout/types";
 
 export const checkoutRemovePromoCodeUseCase = ({
   checkoutRemovePromoCodeInfra,
